+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Disable EVM mempool due to bug on public nodes broadcast
 - Remove references to time.Now() on release schedule validation
 - Remove unsafe math.Sqrt usage in oracle ballot standard deviation calculation
+- Remove panic on oracle slash logic when validator is not found
 
 ## v6.0.0 - 2025-11-25